+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Add Ubuntu 22.04 LTS support.
 - Add preliminary ARM64 image support for T2A instances.
 - Add Rocky Linux 8 support.
+- Adjust logging for config not found.
 
 ## \[5.6.2\]